src/app/core/components/pokemon-details/pokemon-details.component.spec.ts
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { PokemonDetailsComponent } from './pokemon-details.component';

describe('PokemonDetailsComponent', () => {
  let component: PokemonDetailsComponent;
  let fixture: ComponentFixture<PokemonDetailsComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ PokemonDetailsComponent ],
      imports: [HttpClientTestingModule]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(PokemonDetailsComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should Input Id Property', () => {
    const id = 1;
    component.id = id;
    expect(component.id).toBe(id);
  });

  it('should emit colorType', () => {
    const colorType = '#fff';
    component.colorType.subscribe(value => {
      expect(value).toBe(colorType);
    });
    component.colorType.emit(colorType);
  });
});
